Add "hw_vif_multiqueue_enabled: true" property on nfvbenchvm images
Create loop VM and generator VM images in OpenStack with the property
"hw_vif_multiqueue_enabled: true".  Virtio interfaces of virtual machines using
those images will then be created with a number of queues equal to the number
of vcpus, with a maximum value of 8 queues per interface.

The property has no effect on SR-IOV or PCI passthrough network interfaces, but
it is useful for virtio interfaces.

Remark: when nfvbench deploys a loop VM using an image that does not have the
"hw_vif_multiqueue_enabled: true" property set, it will try to set the
property.  But this will fail if the OpenStack user configured for nfvbench
does not have the rights to do so.

Signed-off-by: Gwenael Lambrouin <gwenael.lambrouin@orange.com>