Set missing env variables for TestAPI
In order to have a comprehensive setting of TestAPI result fields for
the nfvbench results written by xtesting, define environment variables
for the missing fields:

- PROJECT_NAME: name of the test project (defaults to "nfvbench")
- INSTALLER_TYPE: identifier for the IaaS installation (defaults to
  "unknown")
- DEPLOY_SCENARIO: identifier for the test scenario (defaults to
  "os-nosdn-nofeature-noha")

It is up to the caller of the nfvbench Ansible role to set these
environment variables.  When they are not defined, the default values
will be used.