Fix results archive not created
When the BUILD_TAG environment variable contains a slash, the test
results archive creation fails.

We fix this by removing all "special" characters found in the BUILD_TAG
and replacing them with an underscore.  In the end, we only keep
unchanged the alphanumerical characters (A-Za-z0-9), the dash (-) and
the underscore (_).