Fix some undetected test failures
Before that patch, the detection of failures occuring during the
execution of the xtesting run_tests command was only based on the
recognition of some strings in the standard output.  But sometimes this
does not work.  For instance, if run_tests is killed by the operating
system because of a badly written test containing an infinite recursion,
it is likely that no string indicating a failure will be written to
stdout.

This is now fixed with the addition of a new failure condition: fail
when run_tests returns a non-zero code.

Remark: some of the other failure conditions are preserved for the
corner case were the tests are not run with the behave framework but are
run directly with nfvbench instead.  They are necessary because nfvbench
will not always return a non-zero code in case of failure (for instance:
nfvbench returns 0 when the end-to-end connectivity test fails).