Add OpenStack API early check
Try to authenticate to the OpenStack cloud at an early stage so that
authentication issues appear clearly in the log.

Before, this check was performed at the same time as the "Check user is
admin" task, but the results were confusing: we could believe that a
failure of the "Check user is admin" task indicated that the user was
not admin while it actually indicated an API authentication issue.

Now, the two checks are done separately.  Besides, the "Check user is
admin" is not done anymore when the user knows he is not an admin and
sets the admin Ansible variable to false.