Revert "Improve trafic generator performance"
This reverts commit e70fe45877cef41e082f49176ee222ad474aea16.

Although nfvbench cache_size parameter can increase the rate of packets
generated by TRex, further testing showed that this parameter has some
undesired effects on the nature of the generated traffic: when
cache_size is lower than the requested flow count, the amount of traffic
for each flow count is not correctly balanced.  This is generally not
acceptable when the system under test is sensible to the number of
flows.

Because of that, we revert the default config of the cache_size and
we use nfvbench default: cache_size=0.

Also, the same tests showed that setting no_flow_stats to true did not
impact significantly the generated traffic rate, so we also stick to
nfvbench default: no_flow_stats=false

Remark: this patch set reverts the default configuration of
ansible-nfvbench-role, but it is still possible to override those
parameters using Ansible extra vars.